*: add crossout

+++ app/src/main/java/pl/edu/pjwstk/s999844/shoppinglist/activities/MainActivity.kt
@@ -96,7 +96,7 @@ class MainActivity : AbstractShoppingActivity() {
 			?: return
 
 		val newAmount = change?.plus(dbItem.amount)
-		if (newAmount == null || newAmount <= 0) {
+		if (newAmount == null || (!this.settings.crossoutActive && newAmount <= 0)) {
 			shoppingListDao.delete(dbItem)
 		} else {
 			dbItem.amount = newAmount
